+++ libraries/botbuilder-dialogs/src/prompts/adaptiveCardPrompt.ts
